+++ src/Calculator/CalculatorInterface.php
@@ -2,6 +2,7 @@
 
 namespace Ayeo\Price\Calculator;
 
+use Ayeo\Price\Money;
 use Ayeo\Price\Price;
 use LogicException;
 
@@ -15,7 +16,13 @@ interface CalculatorInterface
      * @throws LogicException
      */
     public function subtract(Price $left, Price $right): Price;
+    /**
+     * @throws LogicException
+     */
     public function multiply(Price $left, float $times): Price;
+    /**
+     * @throws LogicException
+     */
     public function divide(Price $left, float $times): Price;
-    public function hasSameCurrencies(Price $left, Price $right): bool;
+    public function decorateMoney(Money $money): Money;
 }