+++ src/Decorator/Money/DecoratorInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Ayeo\Price\Decorator\Money;
+
+use Ayeo\Price\Money;
+
+interface DecoratorInterface
+{
+    public function decorateMoney(Money $money): Money;
+}