+++ changelog.md
@@ -1,5 +1,9 @@
 This is draft, will polish it later
 
+1.1.1
+- Add calculators for math operations
+- Add money decorators
+
 1.1.0
 - Updated code to php 7.4 standards